feat: Add adapter for tableview
