+++ ImageFeed/Screens/ImagesList/ImagesListViewController.swift
@@ -10,14 +10,11 @@ import UIKit
 final class ImagesListViewController: UIViewController {
         
     private let tableView = UITableView()
-    private let dataSource: ImagesListDataSource = .init(pictures: Picture.pictures)
-    
-    private var didAnimateCells: [IndexPath: Bool] = [:]
+    private var adapter: ImagesListTableViewAdapter?
     
     private let refreshControl = UIRefreshControl()
     private var hasRefreshed = false {
         didSet {
-            didAnimateCells = [:]
             reloadView()
         }
     }
@@ -33,19 +30,12 @@ final class ImagesListViewController: UIViewController {
 
 private extension ImagesListViewController {
     private func setup() {
-        setupTableView()
-        setupRefreshControl()
-    }
-    
-    func setupTableView() {
-        tableView.delegate = self
-        tableView.dataSource = dataSource
+        adapter = ImagesListTableViewAdapter(
+            tableView,
+            self,
+            PictureViewModel.pictureViewModels
+        )
         
-        tableView.register(ImagesListCell.self, forCellReuseIdentifier: ImagesListCell.reuseID)
-    }
-    
-    func setupRefreshControl() {
-        refreshControl.tintColor = Theme.color(usage: .ypWhite)
         refreshControl.addTarget(self, action: #selector(refreshContent), for: .valueChanged)
         tableView.refreshControl = refreshControl
     }
@@ -57,6 +47,8 @@ private extension ImagesListViewController {
         tableView.separatorStyle = .none
         
         tableView.contentInset = Theme.contentInset(kind: .table)
+        
+        refreshControl.tintColor = Theme.color(usage: .ypWhite)
     }
     
     func applyLayout() {
@@ -75,40 +67,19 @@ private extension ImagesListViewController {
 // MARK: - Actions
 private extension ImagesListViewController {
     @objc func refreshContent() {
-        dataSource.shufflePictures()
+        adapter?.shufflePictures()
         hasRefreshed.toggle()
     }
     
     func reloadView() {
         tableView.refreshControl?.endRefreshing()
-        DispatchQueue.main.asyncAfter(deadline: .now() + 0.5) { [weak self] in
-            guard let self = self else { return }
-            self.tableView.performBatchUpdates {
-                self.tableView.reloadSections([0], with: .automatic)
-            }
-        }
+        adapter?.reloadView()
     }
 }
 
-// MARK: - UITableViewDelegate
-extension ImagesListViewController: UITableViewDelegate {
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        tableView.deselectRow(at: indexPath, animated: true)
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        dataSource.getCellHeight(indexPath)
-    }
-    
-    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        guard didAnimateCells[indexPath] == nil else { return }
-        didAnimateCells[indexPath] = true
-        
-        cell.transform3DMakeRotation(
-            degree: 90,
-            x: 0, y: 1, z: 0,
-            duration: 0.85,
-            delay: 0.1
-        )
+// MARK: - ImagesListTableViewAdapterDelegate
+extension ImagesListViewController: ImagesListTableViewAdapterDelegate {
+    func didSelectImage(_ adapter: ImagesListTableViewAdapter, didSelect item: PictureViewModel) {
+        print(item)
     }
 }
